Creating example template for starters
This update shows how starter project will look moving forward.

It starts with a basic homepage with next steps. The application
is only one component, with no additional services or dependencies
(out of the box). Hot loading is also enabled.

For this example, I also showed what it would look like to add in
an optional feature into a starter project. I've commented out
everything that relates to the feature, and added in a section
under the README for "Additional Learning" with steps on how
to enable this feature.